move util func getUsers to common
