+++ graphql/appointment/fields.ts
@@ -6,12 +6,12 @@ import { getSessionStudent, getUserForSession, isElevated, isSessionStudent } fr
 import { Deprecated } from '../util';
 import { LimitEstimated } from '../complexity';
 import { prisma } from '../../common/prisma';
-import { getUserIdTypeORM, getUserTypeAndIdForUserId } from '../../common/user';
+import { getUserIdTypeORM, getUserTypeAndIdForUserId, getUsers } from '../../common/user';
 
 @ObjectType()
 class AppointmentParticipant {
     @Field((_type) => String, { nullable: true })
-    userId: string;
+    userID: string;
     @Field((_type) => Int, { nullable: true })
     id: number;
     @Field((_type) => String, { nullable: true })
@@ -29,15 +29,13 @@ class AppointmentParticipant {
 @ObjectType()
 class Organizer {
     @Field((_type) => String, { nullable: true })
-    userId: string;
+    userID: string;
     @Field((_type) => Int)
     id: number;
     @Field((_type) => String)
     firstname: string;
     @Field((_type) => String)
     lastname: string;
-    @Field((_type) => Boolean)
-    isStudent: boolean;
 }
 
 @Resolver((of) => Appointment)
@@ -78,77 +76,14 @@ export class ExtendedFieldsLectureResolver {
     @Authorized(Role.OWNER, Role.APPOINTMENT_PARTICIPANT)
     @LimitEstimated(30)
     async participants(@Root() appointment: Appointment, @Arg('take', (type) => Int) take: number, @Arg('skip', (type) => Int) skip: number) {
-        const studentIds = [];
-        const pupilIds = [];
-        appointment.participantIds.forEach((userId) => {
-            const [type, id] = getUserTypeAndIdForUserId(userId);
-            if (type === 'pupil') {
-                pupilIds.push(id);
-            }
-            if (type === 'student') {
-                studentIds.push(id);
-            }
-        });
-        const studentParticipants = (
-            await prisma.student.findMany({
-                where: {
-                    id: {
-                        in: studentIds,
-                    },
-                },
-                take,
-                skip,
-                select: {
-                    id: true,
-                    firstname: true,
-                    lastname: true,
-                    isStudent: true,
-                },
-            })
-        ).map((p) => ({ ...p, isStudent: true, userId: getUserIdTypeORM(p) }));
-
-        const pupilParticipants = (
-            await prisma.pupil.findMany({
-                where: {
-                    id: {
-                        in: pupilIds,
-                    },
-                },
-                take,
-                skip,
-                select: {
-                    id: true,
-                    firstname: true,
-                    lastname: true,
-                    isPupil: true,
-                },
-            })
-        ).map((p) => ({ ...p, isPupil: true, userId: getUserIdTypeORM(p) }));
-        return [...studentParticipants, ...pupilParticipants];
+        return (await getUsers(appointment.participantIds)).map(({ email, ...rest }) => ({ ...rest }));
     }
 
     @FieldResolver((returns) => [Organizer])
     @Authorized(Role.USER)
     @LimitEstimated(5)
     async organizers(@Root() appointment: Appointment, @Arg('take', (type) => Int) take: number, @Arg('skip', (type) => Int) skip: number) {
-        const organizerStudentIds = appointment.organizerIds.map((userId) => parseInt(userId.split('/')[1]));
-        return (
-            await prisma.student.findMany({
-                where: {
-                    id: {
-                        in: organizerStudentIds,
-                    },
-                },
-                take,
-                skip,
-                select: {
-                    id: true,
-                    firstname: true,
-                    lastname: true,
-                    isStudent: true,
-                },
-            })
-        ).map((p) => ({ ...p, isStudent: true, userId: getUserIdTypeORM(p) }));
+        return (await getUsers(appointment.organizerIds)).map(({ email, ...rest }) => ({ ...rest }));
     }
 
     // TODO add declinedBy FieldResolver
